Refactor scheduler and message sending in r2d2wm-bot
The internal implementation of the Scheduler struct has been revised to simplify and clarify the creation of asynchronous tasks. This involved removing explicit clonings and structuring the sending of messages as a separate function 'send_message_async'. Additionally, the 'send' function in the 'ScheduledMessage' struct has been renamed to 'send_to_discord' for better clarity.